# [Profiling] Speed up processing of stacktraces