--- a/docs/changelog/104674.yaml
+++ b/docs/changelog/104674.yaml
@@ -1,0 +1,5 @@
+pr: 104674
+summary: "[Profiling] Speed up processing of stacktraces"
+area: Application
+type: enhancement
+issues: []