So far we kept track of stack frame and executable ids that are associated to stack traces in a sorted collection. The rationale behind this was that the mget should be faster. However, we have seen in benchmarks that the contrary is the case: Not the mgets are the bottleneck but rather insertion is so slow that it defers further processing internally by several hundred milliseconds. When we change to an unsorted (and pre-sized) collection we reduce median response time from ~ 1200ms to 1000ms (or 17%).