@danielmitterdorfer 2024-01-24
Here is a comparison of the results:

|                      Metric |       Baseline |      This PR |   Unit |
|----------------------------:|---------------:|-------------:|-------:|
|     50th percentile latency | 1221.2         | 1006.66      |     ms |
|     90th percentile latency | 1264.46        | 1048.6       |     ms |
|     99th percentile latency | 1297.59        | 1066.28      |     ms |
|    100th percentile latency | 1297.59        | 1080.76      |     ms |

I've also captured an execution trace for both scenarios.

Here is the baseline (the highlighted chunk took 25ms); we can see how progress is deferred by chunk processing long after the last mget has finished:

![baseline_trace](https://github.com/elastic/elasticsearch/assets/1699576/e356d87d-b1d0-4e3e-9f7c-9f1eb5a4ec9e)

For comparison, here is the behavior with this PR (the highlighted chunk took 10ms):

![contender_trace](https://github.com/elastic/elasticsearch/assets/1699576/d29fffe8-369b-41b0-b152-30e5e5da12d5)