@danielmitterdorfer 2024-01-24
> I am curious (again =)): Did you measure the performance of either one of those? Especially using a pre-allocated ordered Set would be interesting.

I did! The number you see above include both changes. The numbers without pre-allocation are not distinguishable from measurement noise. Nevertheless, I wanted to make sure we avoid additional work due to collection resizing.